Adjust excluded files section of configuration.
Exclude list:
* Add package.json and package-lock.json, NPM configuration files.
* Add screenshot.png, which is only referenced by the already-excluded
README.md.

Remove the keep_files section of the configuration. It's purpose is to
list files that should not be clobbered when the _site directory is
being rebuilt. It currently only contains "README.md", and this file
is already explicitly excluded from the build by the exclude section, so
this configuration directive has no effect.