Duplicated hover styles to focus
This supports keyboard-only users by replicating the hover states as focus styles. The catch is that the hover states are on the container _and_ the interactive control, necessitating `:focus-within`. Because this selector will kill an entire style for IE users when added into another selector, it warrants its own stand-alone selector (and still leaves IE users unsupported for focus styles). Also, I do not know SASS. So yeah.