fix #147, make element focus more obvious
Prior to this commit, depending on the browser being used, it may
have been quite difficult to see which element was in focus when
navigating the page using the tab key. This has been corrected by
adjusting the `:focus` pseudo-class styles for selectable elements
so that it no longer depends on inconsistent user-agent styles.

Notes:
    https://user-images.githubusercontent.com/17770407/87440188-b38cdc80-c5bf-11ea-86e3-8b1c8652e25b.png
    https://user-images.githubusercontent.com/17770407/87440518-141c1980-c5c0-11ea-964f-70ff56523438.png

Fixes: https://github.com/tc39/tc39.github.io/issues/147